Rename and update sample project for OpenId
