+++ Swashbuckle.AspNetCore.sln
@@ -91,7 +91,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Swashbuckle.AspNetCore.Test
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSwagClientExample", "test\WebSites\NswagClientExample\NSwagClientExample.csproj", "{9840E751-5845-431C-943B-C75CAE596A45}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenApiIntegration", "test\WebSites\OpenApiIntegration\OpenApiIntegration.csproj", "{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenIdIntegration", "test\WebSites\OpenIdIntegration\OpenIdIntegration.csproj", "{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -475,18 +475,18 @@ Global
 		{9840E751-5845-431C-943B-C75CAE596A45}.Release|x64.Build.0 = Release|Any CPU
 		{9840E751-5845-431C-943B-C75CAE596A45}.Release|x86.ActiveCfg = Release|Any CPU
 		{9840E751-5845-431C-943B-C75CAE596A45}.Release|x86.Build.0 = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|x64.Build.0 = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Debug|x86.Build.0 = Debug|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|x64.ActiveCfg = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|x64.Build.0 = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|x86.ActiveCfg = Release|Any CPU
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2}.Release|x86.Build.0 = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|x64.Build.0 = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Debug|x86.Build.0 = Debug|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|x64.ActiveCfg = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|x64.Build.0 = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|x86.ActiveCfg = Release|Any CPU
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -524,7 +524,7 @@ Global
 		{322813C4-0458-4F98-965D-568AD2C819CE} = {245144DE-BC89-4822-B044-020458BFECC0}
 		{6E3C0128-931F-4438-AEDD-984E0E2B2C7B} = {1669F896-133C-4996-B58C-E7CDA299ADFF}
 		{9840E751-5845-431C-943B-C75CAE596A45} = {245144DE-BC89-4822-B044-020458BFECC0}
-		{5B90D853-BDAA-4641-A0FE-A41AC1A52BB2} = {245144DE-BC89-4822-B044-020458BFECC0}
+		{494211DB-1A5F-4527-AA0C-B6AF4788A1BE} = {245144DE-BC89-4822-B044-020458BFECC0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {36FC6A67-247D-4149-8EDD-79FFD1A75F51}
