+++ test/WebSites/OpenIdIntegration/AuthServer/Views/Consent.cshtml
@@ -1,4 +1,4 @@
-﻿@model OpenApiIntegration.AuthServer.Controllers.ConsentViewModel
+﻿@model OpenIdIntegration.AuthServer.Controllers.ConsentViewModel
 
 <style>
 </style>
