+++ test/WebSites/OpenIdIntegration/AuthServer/Views/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model OpenApiIntegration.AuthServer.Controllers.LoginViewModel
+﻿@model OpenIdIntegration.AuthServer.Controllers.LoginViewModel
 
 <style>
     input {
