+++ test/WebSites/OpenIdIntegration/Properties/launchSettings.json
